Rewrite conditional handler.
The theory is that while our DFA will arrive at a single endpoint, the meaning of that endpoint depends on the start states. Rather than construct a more complicated DFA, we construct a switch statement method that, when a DFA state is reached that has a conditional interpretation, calls the switch statement to determine the correct reduction rule. That is, we generate a conditionalAction: method which handles the conditional rules (including ^ and $).