Sort constants into alphabetical order.
The intent here is that if you are using mutiple grammars in the same code, if you define the same %token declarations in your .y grammar, you will get the same constants and constant values in the generated .h file. This allows multiple grammars to be used side-by-side if you're careful.