Changes to create a closer tie between lex and yacc
The problem with building a re-entrant lexer is trying to get the union tag to work correctly between C++ lexer and C++ yacc tool. We were able to tie the Objective-C code together by having the value field be a generic Objective-C object. But for C++ we want to have a union declaration. (This makes automatic release of objects problematic as the LR(1) stack grows and shrinks, but we assume if we're building a yacc-like tool we can deal with this through a separate allocation pool.)

But this means tying the two halves together.

So this introduces a %union extension on both the lex and yacc side. This also introduces the %lex extension to define the class (and optionally the include file with that class.)

If using the yacc tool as a standalone tool, this generates an interface class OCLexInput class your lexer needs to derive from. If using the lex tool as a standalone tool, the union parameter can be used to create a 'value' field.

If using the yacc tool with the lex tool, you need to include the yacc header file in the lex tool (in the %{ %} header), then define the lex class name in the yacc file using %lex class_name "header_name".

The union of types must then be defined in the yacc file; this gets included in the lex header (via the %{ %} include above), and the presence of the header causes the union to be compiled into the lexer object.

To pass a value up from the lexer you must explicitly add code.

This also alters the %type <x> value; instead of being Objective-C classes, this becomes the union fields. So if you declare %union { int a; }, then the %type is <a>. Unlike the Objective-C generator, types in the union are passed by value rather than adding an implicit pointer.