Move the headers.
So in order to make tight linking of yacc and lex work for C++, we need to actually include the headers specified in the %{...%} fields into the .h file. This allows the requirement to #include the yacc file to get the %union declaration within the lex file to work actually, well, work.