Improve error reporting.
Added code to note the location of the last read token when an error takes place. This is the reported error position on the next call to the error handling method, unless a reduction takes place. (The theory is that this is the correct location to report when the user calls errorWithCode:, in the event an 'error' token is processed, or when an error is caught internally. However, if a reduce is performed, then the correct error location is whatever is the position at the top of the stack--so, for example, an assignment error would be associated with the start of the assignment rule.)